semantic-integrity-analysis-main/requirements.txt
pdfplumber
python-docx
spacy
sentence-transformers
faiss-cpu
numpy
streamlit
transformers
torch
huggingface_hub
reportlab
flask
flask_cors